requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
mysqlclient==2.2.0
redis==5.0.1
celery==5.3.6
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
loguru==0.7.2
python-json-logger==2.0.7
apscheduler==3.10.4
requests==2.31.0 
aioredis~=2.0.1
starlette~=0.27.0
attrs~=25.3.0
redis==5.0.1
asyncmy>=0.2.10
aiomysql>=0.0.21
celery-aio-pool==0.1.0rc8
